+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.3.4-beta.2 - 2022-04-10
+### Updated
+- Updated database field `elementrelations.relations` from `text` to `mediumText`.
+
+### Fixed
+- Fixed performance when redactor/linkit is enabled.
+
 ## 1.3.4-beta.1 - 2022-04-04
 ### Updated
 - Minimize queue job duplicates.